backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.3.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==46.0.3
distro==1.9.0
dj-database-url==3.1.0
Django==6.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==25.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.47.0
google-generativeai==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
numpy==1.26.4
openai==1.3.0
packaging==26.0
pandas==2.1.3
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.3.2
whisper==1.1.10
whitenoise==6.11.0
